Modified code for testing on my Lenovo T420's proper gpu drivers (Win7)
* Modified code for testing on my Lenovo T420's 'proper' gpu drivers
with Windows 7. It doesn't seem to support anything higher than OpenGL
3.0, and the context-creation code(lwjgl-swt) requires at least 3.2 when
specifying the major and minor version, so I had to use zero for
"unspecified".
Things are still a bit buggy, but it's an outdated laptop, so what can
you do? (e.g. setting vsync on one window but turning it off on the
other doesn't work properly; they both share the same vsync setting [are
they technically using the same OpenGL context?])